Fix casing issue with lemma representation
In the common lemma representation of the Generalized and Suffix
transformation classes, the code type has been changed from integer to
long.

The order of the tag and casing has been switched because the tag's
length is longer than expected. The old structure could represent only
0-999 different tags safely. In long it's possible to store 922337203686
different tag.

The constructur and the getter functions has been reworked to fit the
modified type and structure.

Minor modifications has been made with casing too. From now only the
first letter of lemma and token will lowered before comparison.

+++ src/main/java/hu/ppke/itk/nlpg/purepos/common/lemma/GeneralizedLemmaTransformation.java
@@ -74,7 +74,7 @@ public class GeneralizedLemmaTransformation extends
 
 
 	@Override
-	protected Pair<String, Integer> specifyParameters(String word, String lemma, Integer tag, Integer casing) {
+	protected Pair<String, Long> specifyParameters(String word, String lemma, Integer tag, Integer casing) {
 		int removeStart;
 		int removeEnd;
 		String addStart = "";
@@ -94,7 +94,7 @@ public class GeneralizedLemmaTransformation extends
 			addEnd = lemma.substring(posLemma_Word.getLeft()+posLemma_Word.getRight());
 		}
 
-		int code = createCode(tag,casing,removeStart,removeEnd,addEnd.length());
+		long code = createCode(tag,casing,removeStart,removeEnd,addEnd.length());
 		String lemmaStuff = addStart+addEnd;
 		return Pair.of(lemmaStuff, code);
 	}
